+++ dgraph-101/README.md
@@ -1,0 +1,9 @@
+# dgraph with modus 101
+
+Demo from 11/16/24 ModusHack Webinar.
+
+## Dgraph
+
+Start a local instance using Docker.
+
+
